Add files via upload
assumptions2 gives the option of not printing graphs but instead getting an RMSE statistic for each. This is because graphs are often not good fits and take a long time to produce. Check_assumptions_beta_SJstats is from github and is used within the trafo package but no longer exists in the CRAN version of SJstats, so I made a copy of their github beta version.